style: Changed file names to be better described
