+++ src/auth/auth.module.ts
@@ -5,8 +5,8 @@ import { PassportModule } from '@nestjs/passport';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { JwtModule } from '@nestjs/jwt';
 import { AuthController } from './auth.controller';
-import { LocalStrategy } from './strategies/local.strategy';
-import { JwtStrategy } from './strategies/jwt.strategy';
+import { LocalStrategy } from './strategies/auth-user.strategy';
+import { JwtStrategy } from './strategies/jwt-customer.strategy';
 
 @Module({
   imports: [
