+++ src/auth/strategies/jwt-customer.strategy.ts
