+++ src/customer-auth/customer-auth.module.ts
@@ -5,7 +5,7 @@ import { ConfigModule, ConfigService } from '@nestjs/config';
 import { JwtModule } from '@nestjs/jwt';
 import { CustomerAuthController } from './customer-auth.controller';
 import { AuthCustomerStrategy } from './strategies/auth-customer.strategy';
-import { JwtStrategy } from './strategies/jwt.strategy';
+import { JwtStrategy } from './strategies/jwt-customer.strategy';
 import { PassportModule } from '@nestjs/passport';
 
 @Module({
