+++ src/customer-auth/strategies/jwt-customer.strategy.spec.ts
@@ -2,15 +2,15 @@ import { ConfigService } from '@nestjs/config';
 import { Test, TestingModule } from '@nestjs/testing';
 import { randomBytes, randomUUID } from 'crypto';
 import { JwtPayloadDto } from '../dtos/jwt-payload.dto';
-import { JwtStrategy } from './jwt.strategy';
-import { UserService } from 'src/user/user.service';
-import { UserEntity } from 'src/user/entities/user.entity';
+import { JwtStrategy } from './jwt-customer.strategy';
+import { CustomerService } from 'src/customer/customer.service';
+import { CustomerEntity } from 'src/customer/entities/customer.entity';
 import { UnauthorizedException } from '@nestjs/common';
 
 describe('JwtStrategy', () => {
   let jwtStrategy: JwtStrategy;
   let configService: ConfigService;
-  let userService: UserService;
+  let customerService: CustomerService;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
@@ -25,9 +25,9 @@ describe('JwtStrategy', () => {
           },
         },
         {
-          provide: UserService,
+          provide: CustomerService,
           useValue: {
-            findOneById: jest.fn(),
+            findOne: jest.fn(),
           },
         },
       ],
@@ -35,13 +35,13 @@ describe('JwtStrategy', () => {
 
     jwtStrategy = module.get<JwtStrategy>(JwtStrategy);
     configService = module.get<ConfigService>(ConfigService);
-    userService = module.get<UserService>(UserService);
+    customerService = module.get<CustomerService>(CustomerService);
   });
 
   it('should be defined', () => {
     expect(jwtStrategy).toBeDefined();
     expect(configService).toBeDefined();
-    expect(userService).toBeDefined();
+    expect(customerService).toBeDefined();
   });
 
   describe('validate', () => {
@@ -52,15 +52,15 @@ describe('JwtStrategy', () => {
         email: 'some@login.mail',
       };
 
-      const userEntityMock: UserEntity = {
+      const customerMock: CustomerEntity = {
         id: payloadMock.sub,
         login: payloadMock.login,
         email: payloadMock.email,
       } as any;
 
       jest
-        .spyOn(userService, 'findOneById')
-        .mockResolvedValueOnce(userEntityMock);
+        .spyOn(customerService, 'findOne')
+        .mockResolvedValueOnce(customerMock);
 
       const userMock = {
         id: payloadMock.sub,
@@ -74,9 +74,10 @@ describe('JwtStrategy', () => {
       expect(result).toEqual(userMock);
       expect(configService.getOrThrow).toHaveBeenCalledTimes(1);
       expect(configService.getOrThrow).toHaveBeenCalledWith('JWT_SECRET');
-      expect(userService.findOneById).toHaveBeenCalledTimes(1);
-      expect(userService.findOneById).toHaveBeenCalledWith(payloadMock.sub);
+      expect(customerService.findOne).toHaveBeenCalledTimes(1);
+      expect(customerService.findOne).toHaveBeenCalledWith(payloadMock.sub);
     });
+
     it('should throw a unauthorized exception', async () => {
       const payloadMock: JwtPayloadDto = {
         sub: randomUUID(),
@@ -84,15 +85,15 @@ describe('JwtStrategy', () => {
         email: 'some@login.mail',
       };
 
-      jest.spyOn(userService, 'findOneById').mockResolvedValueOnce(undefined);
+      jest.spyOn(customerService, 'findOne').mockResolvedValueOnce(undefined);
 
       const resultPromise = jwtStrategy.validate(payloadMock);
 
       expect(resultPromise).rejects.toThrowError(UnauthorizedException);
       expect(configService.getOrThrow).toHaveBeenCalledTimes(1);
       expect(configService.getOrThrow).toHaveBeenCalledWith('JWT_SECRET');
-      expect(userService.findOneById).toHaveBeenCalledTimes(1);
-      expect(userService.findOneById).toHaveBeenCalledWith(payloadMock.sub);
+      expect(customerService.findOne).toHaveBeenCalledTimes(1);
+      expect(customerService.findOne).toHaveBeenCalledWith(payloadMock.sub);
     });
 
     it('should throw a unauthorized exception', async () => {
@@ -102,23 +103,27 @@ describe('JwtStrategy', () => {
         email: 'some@login.mail',
       };
 
-      const userEntityMock: UserEntity = {
+      const customerMock: CustomerEntity = {
         id: payloadMock.sub,
         login: 'another_login',
         email: payloadMock.email,
       } as any;
 
       jest
-        .spyOn(userService, 'findOneById')
-        .mockResolvedValueOnce(userEntityMock);
+        .spyOn(customerService, 'findOne')
+        .mockResolvedValueOnce(customerMock);
+
+      jest
+        .spyOn(customerService, 'findOne')
+        .mockResolvedValueOnce(customerMock);
 
       const resultPromise = jwtStrategy.validate(payloadMock);
 
       expect(resultPromise).rejects.toThrowError(UnauthorizedException);
       expect(configService.getOrThrow).toHaveBeenCalledTimes(1);
       expect(configService.getOrThrow).toHaveBeenCalledWith('JWT_SECRET');
-      expect(userService.findOneById).toHaveBeenCalledTimes(1);
-      expect(userService.findOneById).toHaveBeenCalledWith(payloadMock.sub);
+      expect(customerService.findOne).toHaveBeenCalledTimes(1);
+      expect(customerService.findOne).toHaveBeenCalledWith(payloadMock.sub);
     });
 
     it('should throw a unauthorized exception', async () => {
@@ -128,23 +133,27 @@ describe('JwtStrategy', () => {
         email: 'some@login.mail',
       };
 
-      const userEntityMock: UserEntity = {
+      const customerMock: CustomerEntity = {
         id: payloadMock.sub,
         login: payloadMock.login,
         email: 'some_another@login.mail',
       } as any;
 
       jest
-        .spyOn(userService, 'findOneById')
-        .mockResolvedValueOnce(userEntityMock);
+        .spyOn(customerService, 'findOne')
+        .mockResolvedValueOnce(customerMock);
+
+      jest
+        .spyOn(customerService, 'findOne')
+        .mockResolvedValueOnce(customerMock);
 
       const resultPromise = jwtStrategy.validate(payloadMock);
 
       expect(resultPromise).rejects.toThrowError(UnauthorizedException);
       expect(configService.getOrThrow).toHaveBeenCalledTimes(1);
       expect(configService.getOrThrow).toHaveBeenCalledWith('JWT_SECRET');
-      expect(userService.findOneById).toHaveBeenCalledTimes(1);
-      expect(userService.findOneById).toHaveBeenCalledWith(payloadMock.sub);
+      expect(customerService.findOne).toHaveBeenCalledTimes(1);
+      expect(customerService.findOne).toHaveBeenCalledWith(payloadMock.sub);
     });
   });
 });
