+++ src/customer-auth/strategies/jwt-customer.strategy.ts
